% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvSemAnnualCompSingleCF.R
\name{fvSemiAnnualCompSingleCF}
\alias{fvSemiAnnualCompSingleCF}
\title{Computing Future Value of a Single Cash Flow for Semi-Annual Compounding Frequency.}
\usage{
fvSemiAnnualCompSingleCF(r, n, pv)
}
\arguments{
\item{r}{A number.}

\item{n}{A number.}

\item{pv}{A number.}
}
\value{
Input values to three arguments \code{r} , \code{n} and \code{pv}.
}
\description{
Computing Future Value of a Single Cash Flow for Semi-Annual Compounding Frequency.
}
\details{
The method \code{fvSemiAnnualCompSingleCF()} is developed to Compute Future Value of a Single Cash Flow under Semi-Annual Compounding Frequency.The method gives FVFuture Value when values are passed to its three arguments. Here \code{r} is annual rate, \code{n} is number of years and \code{pv} is Present Value.
}
\examples{
fvSemiAnnualCompSingleCF(0.08,10,2000)
fvSemiAnnualCompSingleCF(0.08,10,-2000)
}
\references{
\insertRef{humm}{tvmComp}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
