% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvSemAnnualCompSingleCF.R
\name{pvSemiAnnualCompSingleCF}
\alias{pvSemiAnnualCompSingleCF}
\title{Computing Present Value of a Single Cash Flow for Semi-Annual Compounding.}
\usage{
pvSemiAnnualCompSingleCF(r, n, fv)
}
\arguments{
\item{r}{A number.}

\item{n}{A number.}

\item{fv}{A number.}
}
\value{
Input values to three arguments \code{r} , \code{n} and \code{fv}.
}
\description{
Computing Present Value of a Single Cash Flow for Semi-Annual Compounding.
}
\details{
The method \code{pvSemiAnnualCompSingleCF()} is developed to compute Present Value of a Single Cash Flow for Semi-Annual Compounding Frequency and gives Present Value when values are passed to its three arguments. Here \code{r} is annual rate, \code{n} is number of years and \code{fv} is Future Value.
}
\examples{
pvSemiAnnualCompSingleCF(0.08,10,2000)
pvSemiAnnualCompSingleCF(0.08,10,-2000)
}
\references{
\insertRef{humm}{tvmComp}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
