% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvSemAnnuallyPaidAnnuityCompSemiAnnually.R
\name{pvSemiAnnuallyPaidAnnuityCompSemiAnnually}
\alias{pvSemiAnnuallyPaidAnnuityCompSemiAnnually}
\title{Computing Present Value of Semi-Annually Paid Ordinary Annuity or Annuity Due that is Compounded Semi-Annually.}
\usage{
pvSemiAnnuallyPaidAnnuityCompSemiAnnually(r, n, pmt, bgn)
}
\arguments{
\item{r}{A number.}

\item{n}{A number.}

\item{pmt}{A number.}

\item{bgn}{A number.}
}
\value{
Input values to four arguments \code{r} , \code{n} , \code{pmt} and \code{bgn}
}
\description{
Computing Present Value of Semi-Annually Paid Ordinary Annuity or Annuity Due that is Compounded Semi-Annually.
}
\details{
The method \code{pvSemiAnnuallyPaidAnnuityCompSemiAnnually()} is developed to compute Present Value of Semi-Annually Paid Ordinary Annuity or Annuity Due that is Compounded Semi-Annually and gives Present Value when values are passed to its four arguments. Here \code{r} is annual rate, \code{n} is number of years, \code{pmt} is amount of one annuity and \code{bgn} is the computational mode. (Enter 1 when annuity payment occurs at the beginning of the period; 0 for end of period payments)
}
\examples{
pvSemiAnnuallyPaidAnnuityCompSemiAnnually(0.06,12, -200, 1)
pvSemiAnnuallyPaidAnnuityCompSemiAnnually(0.06,12, -200, 0)
pvSemiAnnuallyPaidAnnuityCompSemiAnnually(0.06,12, 200, 1)
pvSemiAnnuallyPaidAnnuityCompSemiAnnually(0.06,12, 200, 0)
}
\references{
\insertRef{humm}{tvmComp}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
