% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvVariousCompSinglCF.R
\name{pvVariousCompSingleCF}
\alias{pvVariousCompSingleCF}
\title{Calculates Present Value of Single Cash Flow for various Compounding Frequencies.}
\usage{
pvVariousCompSingleCF(r, n, fv, cfreq)
}
\arguments{
\item{r}{A number.}

\item{n}{A number.}

\item{fv}{A number.}

\item{cfreq}{A number.}
}
\value{
Input values to four arguments \code{r} , \code{n} , \code{fv} and \code{cfreq}
}
\description{
Calculates Present Value of Single Cash Flow for various Compounding Frequencies.
}
\details{
The method \code{pvVariousCompSingleCF()} is developed to calculate Present Value of Single Cash Flow for various Compounding Frequencies and gives Present Value when values are passed to its four arguments. Here \code{r} is annual rate, \code{n} is number of years, \code{pv} is Present Value and \code{cfreq} is Compounding Frequency value that must be either 1 or 2 or 4 or 12 or 365.
}
\examples{
pvVariousCompSingleCF( 0.08,10,2000,2)
pvVariousCompSingleCF( 0.08,10,2000,4)
pvVariousCompSingleCF( 0.08,10,2000,12)
pvVariousCompSingleCF( 0.08,10,2000,365)
}
\references{
\insertRef{humm}{tvmComp}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
